requirements.txt
pika >= 0.11.0
python-dotenv >= 0.7.1
pandas >= 0.20.3
gensim >= 3.0.0
pymongo >= 3.5.1
simplejson >= 3.11.1